quickstart/requirements.txt
khaos-agent>=1.0.0
google-generativeai>=0.5.0
